updated crud transactions
